Disabled CRAM-MD5 at some point.
I think it was because we were advertising CRAM-MD5 support, but then it
would fail to actually work? I can't remember.

We only allow auth over encrypted connections anyhow (I think!), so it doesn't
matter if we get a non-hashed password in that case, I guess.

I don't know. Whatever.